add support for flux vae. ~ wip

+++ src/maxdiffusion/models/modeling_flax_pytorch_utils.py
@@ -243,7 +243,7 @@ def convert_lora_pytorch_state_dict_to_flax(pt_state_dict, params, network_alpha
 
 def convert_pytorch_state_dict_to_flax(pt_state_dict, flax_model, init_key=42):
   # Step 1: Convert pytorch tensor to numpy
-  pt_state_dict = {k: v.numpy() for k, v in pt_state_dict.items()}
+  pt_state_dict = {k: v.float().numpy() for k, v in pt_state_dict.items()}
 
   # Step 2: Since the model is stateless, run eval_shape to get the pytree structure
   random_flax_params = flax_model.init_weights(PRNGKey(init_key), eval_only=True)
